refactor(landing): split the tour scenes into modules on new chrome
Six panels in one 391-line file is the shape that makes the rest of this
redesign serial: two scenes are new, four are being rebuilt, and every
one of them would have been an edit to the same file. Each scene is now
its own module behind one registry, so they can be written independently.

The shared chrome helper stops calling the old sidebar and status
renderers and takes a rail fixture instead, which is what lets the tour's
dead workspace-status export be deleted in the next step. Scenes that
are not about the rail get one compact shared fixture rather than
inventing their own.

Scene bodies move across verbatim. Rewriting them here would have mixed
a structural move with six content changes in one diff; they are each a
task of their own. The two new scenes land as placeholders for the same
reason.

Agent marks now go through the shared renderer, which has the null
branch the old local helper lacked — it threw on any agent without a
brand asset, which is every future one.
