+++ marketing/landing-prototype/src/tour/scenes/usage.js
@@ -1,0 +1,14 @@
+/**
+ * Panel 5 — the token usage dashboard.
+ *
+ * Placeholder: the chrome with an empty stage. The body is Usage → Overview —
+ * the eyebrow, the large figure, the range selector with `All` active, the
+ * estimate note and the two per-agent rows the scanner actually covers — and
+ * it is written by its own task, which owns this file from here on.
+ */
+
+import { frame } from "./chrome.js";
+
+export function usage() {
+  return frame("");
+}
